Server function for UnilSports Shiny Application

This is a function that builds a server for the UnilSports shiny application. This function is then used as the secoond parameter in the UnilSports_gui function.

Usage

build_server(clean_sport_schedule)

Arguments

clean_sport_schedule

The data frame output from the get_cleanschedule_met. The output of this function should not be modified, so this function can apply the integer optimization technique properly.

Value

a shiny server output

Author

Arturo Garcia Luna Beltran Manunpat Sirijaturaporn Nadine Obeid Katti Irastorza

Examples

build_server(clean_sport_schedule)
#> function (input, output, session) 
#> {
#>     observe({
#>         update_date <- input$date
#>         update_schedule <- filter(clean_sport_schedule, Date == 
#>             update_date)
#>         update_time <- input$time
#>         if (!is.null(update_time)) {
#>             update_schedule <- mutate(update_schedule, time = 0)
#>             timetemp <- as.data.frame(matrix(as.numeric(gsub("\\:", 
#>                 ".", unlist(strsplit(sort(update_time), "–")))), 
#>                 ncol = 2, byrow = TRUE))
#>             n <- length(update_time)
#>             selected_time <- data.frame()
#>             for (i in 1:n) {
#>                 if (i == 1) {
#>                   output_temp <- timetemp[i, ]
#>                 }
#>                 else {
#>                   output_temp2 <- timetemp[i, ]
#>                   if (output_temp[2] == output_temp2[1]) {
#>                     output_temp[2] <- output_temp2[2]
#>                   }
#>                   else {
#>                     selected_time <- rbind(selected_time, output_temp)
#>                     output_temp <- timetemp[i, ]
#>                   }
#>                 }
#>                 if (i == n) {
#>                   selected_time <- rbind(selected_time, output_temp)
#>                 }
#>             }
#>             k <- nrow(selected_time)
#>             for (j in 1:k) {
#>                 update_schedule$time <- update_schedule$time | 
#>                   (update_schedule$`Start time` >= selected_time[j, 
#>                     1] & update_schedule$`End time` <= selected_time[j, 
#>                     2])
#>                 update_schedule$time <- 1 * update_schedule$time
#>             }
#>             update_schedule <- update_schedule %>% filter(time == 
#>                 1)
#>         }
#>         update_activity <- unique(update_schedule)$Activity
#>         updateSelectInput(session, "activity", choices = update_activity)
#>     })
#>     optimize_output <- eventReactive(input$opt, {
#>         outputtemp <- optimize_schedule(clean_sport_schedule, 
#>             input$date, input$activity, input$time, input$calburn, 
#>             input$weight, as.numeric(input$no_dup))
#>     })
#>     output$optim_table <- renderTable(hover = T, {
#>         data.frame(optimize_output()[2]) %>% select(-c(table_result.Date, 
#>             table_result.time, table_result.METs, table_result.p)) %>% 
#>             rename(`Start Time` = table_result.Start.time, `End Time` = table_result.End.time, 
#>                 `Duration (min)` = table_result.Duration_min, 
#>                 Activity = table_result.Activity, Location = table_result.Location, 
#>                 Calburn = table_result.calburn)
#>     })
#>     output$optim_result <- renderText({
#>         if (optimize_output()[1] == 1) {
#>             sprintf("The optimization is successful! Total Exercise Duration: %.0f min, Total Calorie Burn: %.0f", 
#>                 optimize_output()[5], optimize_output()[4])
#>         }
#>         else if (optimize_output()[1] == 2) {
#>             sprintf("Cannot reach the calorie burn target. The optimizer returns the solution with the highest possible calorie burn. Total Exercise Duration: %.0f min, Total Calorie Burn: %.0f", 
#>                 optimize_output()[5], optimize_output()[4])
#>         }
#>         else {
#>             sprintf(cat("No feasible solution found"))
#>         }
#>     })
#>     output$calburnplot <- renderPlotly({
#>         optim_plot <- optimize_output()$table_result
#>         p <- pie_optim(optim_plot)
#>         p[[1]]
#>     })
#>     output$durationplot <- renderPlotly({
#>         optim_plot <- optimize_output()$table_result
#>         p <- pie_optim(optim_plot)
#>         p[[2]]
#>     })
#>     output$activity_table <- renderTable({
#>         data.frame(optimize_output()[3]) %>% select(-c(activity_selected.Date, 
#>             activity_selected.time, activity_selected.METs, activity_selected.p)) %>% 
#>             rename(`Start Time` = activity_selected.Start.time, 
#>                 `End Time` = activity_selected.End.time, `Duration (min)` = activity_selected.Duration_min, 
#>                 Activity = activity_selected.Activity, Location = activity_selected.Location, 
#>                 Calburn = activity_selected.calburn)
#>     })
#>     output$image <- renderImage({
#>         list(src = "image.png", align = "right", height = "100%", 
#>             width = "auto", deleteFile = FALSE)
#>     })
#> }
#> <bytecode: 0x56140f300ff8>
#> <environment: 0x56140f2edb90>