User Interface function for UnilSports Shiny Application

This is a function that builds a user interface for the UnilSports shiny application. This function is then used as the first parameter of the UnilSports_gui function.

Usage

build_ui(clean_sport_schedule)

Arguments

clean_sport_schedule

The data frame output from the get_cleanschedule_met. The output of this function should not be modified, so this function can apply the integer optimization technique properly.

Value

a shiny web application dashboard that builds a user interface for the UnilSports shiny application. This function is then used as one of the parameters of the UnilSports_gui function.

Author

Arturo Garcia Luna Beltran Manunpat Sirijaturaporn Nadine Obeid Katti Irastorza

Examples

build_ui(clean_sport_schedule)
#> Warning: Navigation containers expect a collection of `bslib::nav()`/`shiny::tabPanel()`s and/or `bslib::nav_menu()`/`shiny::navbarMenu()`s. Consider using `header` or `footer` if you wish to place content above (or below) every panel's contents.
#> <nav class="navbar navbar-default navbar-static-top" role="navigation">
#>   <div class="container-fluid">
#>     <div class="navbar-header">
#>       <span class="navbar-brand">
#>         <strong>Sports Unil Plan</strong>
#>       </span>
#>     </div>
#>     <ul class="nav navbar-nav" data-tabsetid="7704">
#>       <li>
#>         <a href="#tab-7704-1" data-toggle="tab" data-bs-toggle="tab"></a>
#>       </li>
#>     </ul>
#>   </div>
#> </nav>
#> <div class="container-fluid">
#>   <div class="tab-content" data-tabsetid="7704">
#>     <div class="row" id="tab-7704-1">
#>       <div class="col-sm-4">
#>         <form class="well" role="complementary">
#>           <div class="form-group shiny-input-container">
#>             <label class="control-label" id="date-label" for="date">
#>               <strong>Choose a date you want to workout</strong>
#>             </label>
#>             <div>
#>               <select id="date"><option value="2022-12-14" selected>2022-12-14</option>
#> <option value="2022-12-15">2022-12-15</option>
#> <option value="2022-12-16">2022-12-16</option>
#> <option value="2022-12-17">2022-12-17</option>
#> <option value="2022-12-18">2022-12-18</option>
#> <option value="2022-12-19">2022-12-19</option>
#> <option value="2022-12-20">2022-12-20</option></select>
#>               <script type="application/json" data-for="date" data-nonempty="">{"plugins":["selectize-plugin-a11y"]}</script>
#>             </div>
#>           </div>
#>           <div class="form-group shiny-input-container">
#>             <label class="control-label" id="time-label" for="time">
#>               <strong>Choose time</strong>
#>             </label>
#>             <div>
#>               <select id="time" multiple="multiple"><option value="07:00 – 08:00">07:00 – 08:00</option>
#> <option value="08:00 – 09:00">08:00 – 09:00</option>
#> <option value="09:00 – 10:00">09:00 – 10:00</option>
#> <option value="10:00 – 11:00">10:00 – 11:00</option>
#> <option value="11:00 – 12:00">11:00 – 12:00</option>
#> <option value="12:00 – 13:00">12:00 – 13:00</option>
#> <option value="13:00 – 14:00">13:00 – 14:00</option>
#> <option value="14:00 – 15:00">14:00 – 15:00</option>
#> <option value="15:00 – 16:00">15:00 – 16:00</option>
#> <option value="16:00 – 17:00">16:00 – 17:00</option>
#> <option value="17:00 – 18:00">17:00 – 18:00</option>
#> <option value="18:00 – 19:00">18:00 – 19:00</option>
#> <option value="19:00 – 20:00">19:00 – 20:00</option>
#> <option value="20:00 – 21:00">20:00 – 21:00</option>
#> <option value="21:00 – 22:00">21:00 – 22:00</option>
#> <option value="22:00 – 23:00">22:00 – 23:00</option></select>
#>               <script type="application/json" data-for="time">{"plugins":["selectize-plugin-a11y"]}</script>
#>             </div>
#>           </div>
#>           <div class="form-group shiny-input-container">
#>             <label class="control-label" id="activity-label" for="activity">
#>               <strong>Choose activity</strong>
#>             </label>
#>             <div>
#>               <select id="activity" multiple="multiple"><option value="Réveil musculaire (Cours hybride)">Réveil musculaire (Cours hybride)</option>
#> <option value="Football / Pratique libre">Football / Pratique libre</option>
#> <option value="Basketball / Pratique libre à l&#39;extérieur">Basketball / Pratique libre à l'extérieur</option>
#> <option value="Pilates">Pilates</option>
#> <option value="Pole Sportive / Débutants (Cours 1)">Pole Sportive / Débutants (Cours 1)</option>
#> <option value="Yoga / Avancés (Postures classiques)">Yoga / Avancés (Postures classiques)</option>
#> <option value="Escalade indoor /  Escalade surveillée">Escalade indoor /  Escalade surveillée</option>
#> <option value="Indoor cycling / Cours 08">Indoor cycling / Cours 08</option>
#> <option value="Workout / Total Sculpt">Workout / Total Sculpt</option>
#> <option value="Capoeira / Tous niveaux">Capoeira / Tous niveaux</option>
#> <option value="Football / Jeu dirigé">Football / Jeu dirigé</option>
#> <option value="Danses africaines">Danses africaines</option>
#> <option value="Workout à proximité du CHUV">Workout à proximité du CHUV</option>
#> <option value="Pole Sportive / Intermédiaires (Cours 2)">Pole Sportive / Intermédiaires (Cours 2)</option>
#> <option value="Gymnastique de maintien / Gymnastique de maintien">Gymnastique de maintien / Gymnastique de maintien</option>
#> <option value="Pole Sportive / Débutants (Cours 11 (supplémentaire))">Pole Sportive / Débutants (Cours 11 (supplémentaire))</option>
#> <option value="Musculation connectée / 1. Introduction">Musculation connectée / 1. Introduction</option>
#> <option value="Pole Sportive / Débutants (Cours 3)">Pole Sportive / Débutants (Cours 3)</option>
#> <option value="Pole Sportive / Intermédiaires (Cours 4)">Pole Sportive / Intermédiaires (Cours 4)</option>
#> <option value="Escalade indoor / INITIATION (Automne 2)">Escalade indoor / INITIATION (Automne 2)</option>
#> <option value="Escalade indoor / PERFECTIONNEMENT (Automne 2)">Escalade indoor / PERFECTIONNEMENT (Automne 2)</option>
#> <option value="Mindfulness / Mindfulness">Mindfulness / Mindfulness</option>
#> <option value="Workout / HIIT MACHINE">Workout / HIIT MACHINE</option>
#> <option value="Streetdance / Tous niveaux">Streetdance / Tous niveaux</option>
#> <option value="Basketball / Jeu dirigé">Basketball / Jeu dirigé</option>
#> <option value="Krav Maga / Tous niveaux">Krav Maga / Tous niveaux</option>
#> <option value="Swissball">Swissball</option>
#> <option value="Indoor cycling / Cours 02">Indoor cycling / Cours 02</option>
#> <option value="Rugby à VII / Equipe universitaire de rugby à 7">Rugby à VII / Equipe universitaire de rugby à 7</option>
#> <option value="Escrime / Débutants">Escrime / Débutants</option>
#> <option value="Course à pied / Intermédiaires">Course à pied / Intermédiaires</option>
#> <option value="Volleyball / Jeu">Volleyball / Jeu</option>
#> <option value="Aïkido / Tous niveaux">Aïkido / Tous niveaux</option>
#> <option value="Lindy Hop / Débutants">Lindy Hop / Débutants</option>
#> <option value="Tango argentin / Débutants">Tango argentin / Débutants</option>
#> <option value="Escrime / Avancés et compétiteurs">Escrime / Avancés et compétiteurs</option>
#> <option value="Bachata">Bachata</option>
#> <option value="Lindy Hop / Tous niveaux">Lindy Hop / Tous niveaux</option></select>
#>               <script type="application/json" data-for="activity">{"plugins":["selectize-plugin-a11y"]}</script>
#>             </div>
#>           </div>
#>           <div class="form-group shiny-input-container">
#>             <label class="control-label" id="calburn-label" for="calburn">
#>               <strong>Calories you want to burn today (calories)</strong>
#>             </label>
#>             <input id="calburn" type="number" class="form-control" value="1500" step="50"/>
#>           </div>
#>           <div class="form-group shiny-input-container">
#>             <label class="control-label" id="weight-label" for="weight">
#>               <strong>Your weight (kg)</strong>
#>             </label>
#>             <input id="weight" type="number" class="form-control" value="50"/>
#>           </div>
#>           <div class="form-group shiny-input-container">
#>             <div class="checkbox">
#>               <label>
#>                 <input id="no_dup" type="checkbox"/>
#>                 <span>Do not choose the same activity</span>
#>               </label>
#>             </div>
#>           </div>
#>           <button id="opt" type="button" class="btn btn-default action-button">
#>             <i class="fas fa-dumbbell" role="presentation" aria-label="dumbbell icon"></i>
#>             <strong>  Go!</strong>
#>           </button>
#>         </form>
#>       </div>
#>       <div class="col-sm-8" role="main">
#>         <div id="image" class="shiny-image-output" style="width:auto;height:50px;"></div>
#>         <div class="tabbable">
#>           <ul class="nav nav-tabs" data-tabsetid="7314">
#>             <li class="active">
#>               <a href="#tab-7314-1" data-toggle="tab" data-bs-toggle="tab" data-value="Optimization">Optimization</a>
#>             </li>
#>             <li>
#>               <a href="#tab-7314-2" data-toggle="tab" data-bs-toggle="tab" data-value="Available Activities">Available Activities</a>
#>             </li>
#>           </ul>
#>           <div class="tab-content" data-tabsetid="7314">
#>             <div class="tab-pane active" data-value="Optimization" id="tab-7314-1">
#>               <span class="help-block">Optimization Result</span>
#>               <div id="optim_table" class="shiny-html-output"></div>
#>               <div id="optim_result" class="shiny-text-output"></div>
#>               <div class="row">
#>                 <div class="col-sm-6">
#>                   <br/>
#>                   <br/>
#>                   <div class="plotly html-widget html-widget-output shiny-report-size shiny-report-theme html-fill-item-overflow-hidden html-fill-item" id="calburnplot" style="width:100%;height:350px;"></div>
#>                 </div>
#>                 <div class="col-sm-6">
#>                   <br/>
#>                   <br/>
#>                   <div class="plotly html-widget html-widget-output shiny-report-size shiny-report-theme html-fill-item-overflow-hidden html-fill-item" id="durationplot" style="width:100%;height:350px;"></div>
#>                 </div>
#>               </div>
#>             </div>
#>             <div class="tab-pane" data-value="Available Activities" id="tab-7314-2">
#>               <span class="help-block">List of available activities with respect to the criteria</span>
#>               <div id="activity_table" class="shiny-html-output"></div>
#>             </div>
#>           </div>
#>         </div>
#>       </div>
#>     </div>
#>   </div>
#> </div>